requirements.txt
langchain-classic==1.0.1
langchain-google-genai
langchain
python-dotenv==1.2.1
